feat: moving requestbuilder.tsx from utils to it's own component
